+++ src/main/java/blusunrize/immersiveengineering/client/gui/GuiArcFurnace.java
@@ -8,19 +8,25 @@
 
 package blusunrize.immersiveengineering.client.gui;
 
+import blusunrize.immersiveengineering.api.Lib;
 import blusunrize.immersiveengineering.client.ClientUtils;
+import blusunrize.immersiveengineering.client.gui.elements.GuiButtonIE;
 import blusunrize.immersiveengineering.common.blocks.metal.TileEntityArcFurnace;
 import blusunrize.immersiveengineering.common.blocks.metal.TileEntityMultiblockMetal.MultiblockProcess;
 import blusunrize.immersiveengineering.common.blocks.metal.TileEntityMultiblockMetal.MultiblockProcessInMachine;
 import blusunrize.immersiveengineering.common.gui.ContainerArcFurnace;
+import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.RenderHelper;
+import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.player.InventoryPlayer;
+import net.minecraft.inventory.ClickType;
 
 import java.util.ArrayList;
 
 public class GuiArcFurnace extends GuiIEContainerBase
 {
+	static final String texture = "immersiveengineering:textures/gui/arc_furnace.png";
 	TileEntityArcFurnace tile;
 
 	public GuiArcFurnace(InventoryPlayer inventoryPlayer, TileEntityArcFurnace tile)
@@ -36,7 +42,9 @@ public class GuiArcFurnace extends GuiIEContainerBase
 		super.drawScreen(mx, my, partial);
 		ArrayList<String> tooltip = new ArrayList<String>();
 		if(mx > guiLeft+157&&mx < guiLeft+164&&my > guiTop+22&&my < guiTop+68)
-			tooltip.add(tile.getEnergyStored(null)+"/"+tile.getMaxEnergyStored(null)+" RF");
+			tooltip.add(tile.getEnergyStored(null)+"/"+tile.getMaxEnergyStored(null)+" IF");
+		if(mx > guiLeft+10&&mx < guiLeft+26&&my > guiTop+10&&my < guiTop+26)
+			tooltip.add(I18n.format(Lib.GUI_CONFIG+"arcfurnace.distribute"));
 
 		if(!tooltip.isEmpty())
 		{
@@ -50,7 +58,7 @@ public class GuiArcFurnace extends GuiIEContainerBase
 	protected void drawGuiContainerBackgroundLayer(float f, int mx, int my)
 	{
 		GlStateManager.color(1, 1, 1, 1);
-		ClientUtils.bindTexture("immersiveengineering:textures/gui/arc_furnace.png");
+		ClientUtils.bindTexture(texture);
 		this.drawTexturedModalRect(guiLeft, guiTop, 0, 0, xSize, ySize);
 
 		for(MultiblockProcess process : tile.processQueue)
@@ -71,5 +79,50 @@ public class GuiArcFurnace extends GuiIEContainerBase
 
 		int stored = (int)(46*(tile.getEnergyStored(null)/(float)tile.getMaxEnergyStored(null)));
 		ClientUtils.drawGradientRect(guiLeft+157, guiTop+22+(46-stored), guiLeft+164, guiTop+68, 0xffb51500, 0xff600b00);
+	}
+
+	@Override
+	public void initGui()
+	{
+		super.initGui();
+		this.buttonList.add(new GuiButtonIE(0, guiLeft+10, guiTop+10, 16, 16, null, texture, 179, 0).setHoverOffset(0,16));
+	}
+
+	@Override
+	protected void actionPerformed(GuiButton button)
+	{
+		if(button.id==0)
+			autoSplitStacks();
+	}
+
+	private void autoSplitStacks()
+	{
+		int emptySlot;
+		int largestSlot;
+		int largestCount;
+		do
+		{
+			emptySlot = -1;
+			largestSlot = -1;
+			largestCount = -1;
+			for(int i = 0; i < 12; i++)
+				if(this.inventorySlots.getSlot(i).getHasStack())
+				{
+					int count = this.inventorySlots.getSlot(i).getStack().getCount();
+					if(count > 1&&count > largestCount)
+					{
+						largestSlot = i;
+						largestCount = count;
+					}
+				}
+				else if(emptySlot < 0)
+					emptySlot = i;
+			if(emptySlot>=0 && largestSlot>=0)
+			{
+				this.handleMouseClick(this.inventorySlots.getSlot(largestSlot), largestSlot, 1, ClickType.PICKUP);
+				this.handleMouseClick(this.inventorySlots.getSlot(emptySlot), emptySlot, 0, ClickType.PICKUP);
+			}
+
+		} while(emptySlot >= 0 && largestSlot>=0);
 	}
 }