+++ changelog.md
@@ -1,4 +1,5 @@
 #####Version 0.12-87
+- Added a "Distribute Inputs" button to the ArcFurnace GUI. It keeps splitting the biggest stacks in the input if there is space available (BluSunrize)
 - Changed connectors to accept multiple "bursts" of energy if they (in total) don't exceed the max IO rate (Malte)
 - Changed internal logic of the Assembler to be a lot more straightforward, allowing bucket-like items for fluid ingredients (BluSunrize)
 - Removed clickable JEI areas from the Assembler because they were blocking the output slots and were unnecessary anyway (BluSunrize)